+++ src/feature/CardRegister/components/inputs/PasswordField/PasswordField.stories.tsx
@@ -3,26 +3,27 @@ import type {ChangeEvent} from 'react';
 import type {Meta, StoryObj} from '@storybook/react-vite';
 import {fn} from 'storybook/test';
 
-import CvcField from './CvcField';
+import PasswordField from './PasswordField';
 import type {CardRegisterInputProps} from '../shared.types';
 
 const createInputProps = (value: string): CardRegisterInputProps => ({
+  type: 'password',
   value,
-  maxLength: 3,
-  placeholder: '123',
+  maxLength: 2,
+  placeholder: '',
   onChange: fn(),
   onBlur: fn(),
 });
 
 const meta = {
-  title: 'feature/CardRegister/components/fields/CvcField',
-  component: CvcField,
+  title: 'feature/CardRegister/components/inputs/PasswordField',
+  component: PasswordField,
   tags: ['autodocs'],
   args: {
     inputProps: createInputProps(''),
     errorMessage: '',
   },
-} satisfies Meta<typeof CvcField>;
+} satisfies Meta<typeof PasswordField>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
@@ -31,29 +32,29 @@ export const Empty: Story = {};
 
 export const Filled: Story = {
   args: {
-    inputProps: createInputProps('123'),
+    inputProps: createInputProps('12'),
   },
 };
 
 export const WithErr: Story = {
   args: {
-    inputProps: createInputProps('12'),
-    errorMessage: 'CVC 번호 3자리를 입력해 주세요',
+    inputProps: createInputProps('1'),
+    errorMessage: '비밀번호 앞 2자리를 입력해 주세요',
   },
 };
 
 export const Interactive: Story = {
-  render: function InteractiveCvcField(args) {
-    const [cvcNumber, setCvcNumber] = useState('');
+  render: function InteractivePasswordField(args) {
+    const [password, setPassword] = useState('');
     const inputProps = {
-      ...createInputProps(cvcNumber),
+      ...createInputProps(password),
       onChange: (event: ChangeEvent<HTMLInputElement>) => {
         const nextValue = event.currentTarget.value;
-        if (!/^\d*$/.test(nextValue) || nextValue.length > 3) return;
-        setCvcNumber(nextValue);
+        if (!/^\d*$/.test(nextValue) || nextValue.length > 2) return;
+        setPassword(nextValue);
       },
     };
 
-    return <CvcField {...args} inputProps={inputProps} />;
+    return <PasswordField {...args} inputProps={inputProps} />;
   },
 };
