+++ src/feature/CardRegister/components/inputs/shared.types.ts
